update subscribers new.html.erb and schema to add radius in float

+++ db/migrate/20140912014546_add_column_radius_to_subscribers.rb
@@ -1,5 +1,5 @@
 class AddColumnRadiusToSubscribers < ActiveRecord::Migration
   def change
-    add_column :subscribers, :radius, :integer
+    add_column :subscribers, :radius, :float
   end
 end
